Create Flattened_MOJO
This is the fully flattened source for $MOJO, one of the ARC-20 token used for testing my Solidity contracts. It is located at https://testnet.snowtrace.io/address/0x2b7a4a839ab21bbd8bf158f52980a1e09024e9d2 as $MOJO, the SAQ Governance Token, with 18 decimal places.

A public faucet is not yet available, and is planned for a later version.